add lookupProtocol for windows
This fixes a problem with https://github.com/tinygo-org/net/pull/36
which causes the Windows TinyGO build to fail.  The problem is the
function lookupProtocol was missing for Windows build.  This PR adds the
function for Windows.

I can't really test this because I don't have a Windows system, but by
inspection is seems like it's correct.